# Use PyTorch's DataLoader and enable GPU usage for CLI