--- a/scripts/attacut-cli
+++ b/scripts/attacut-cli
@@ -21,7 +21,6 @@ Options:
 
 import sys
 import os
-from multiprocessing import Pool
 
 import numpy as np
 import torch
@@ -100,7 +99,7 @@ if __name__ == "__main__":
     num_cores = int(arguments["--num-cores"])
     batch_size = int(arguments["--batch-size"])
 
-    assert num_cores >= 1, "Input given to <num-thread> should greather than or equal one"
+    assert num_cores >= 0, "Input given to <num-thread> should greather than or equal one"
 
     if not src:
       print(__doc__)
@@ -126,10 +125,13 @@ if __name__ == "__main__":
 
     total_lines = utils.wc_l(src)
 
-    torch.set_num_threads(num_cores)
-
     device = "cuda" if arguments["--gpu"] else "cpu"
-    print(f"Use {num_cores} cores for processing for {total_lines} lines")
+
+    if num_cores == 0:
+        print(f"Use main process processing for {total_lines} lines")
+    else:
+        print(f"Use {num_cores} cores for processing for {total_lines} lines")
+
     print(f"device={device}")
 
     ds = AttaCutCLIDataset(src, tokenizer, device)