# Benchmark

Current version can be installed using
```
pip install --upgrade git+git://github.com/pythainlp/attacut.git@batch-processing
```

**Test File:** [subset-of-BEST.txt](https://gist.github.com/heytitle/696e908a0d59850e1374c5f823010e3d/raw/02744055414c748e143ca2d862997e59f25a9bcf/gistfile1.txt)
```
# statistics: total lines, characters
wc -l -m subset-of-BEST.txt
13942 2029230 best-val.txt
```

## Results 
Remarks:
- 🥇is the best configuration for each model.
- *Total Execution Time* is the value is the command `time`'s `real`.


### Environment: aws's P2 instance (1 Tesla K80, 4 vCPUs)

|Model| Option | Total Execution Time  | Total Tokenization Time (first → last lines) |  iterations/sec |
|:---:|:--------|:----:|:-------:|:-------:|
|AttaCut-C|--gpu --batch-size=1    --num-cores=1|28s|22s |612| 
||--gpu --batch-size=100 --num-cores=1|15s🥇|9s|1399| 
||--gpu --batch-size=200 --num-cores=1|16s|10s|1308| 
|AttaCut-SC|--gpu --batch-size=1    --num-cores=1|1m18s| 26s | 521| 
||--gpu --batch-size=100 --num-cores=1|58s|7s|1857| 
||--gpu --batch-size=200 --num-cores=1**|58s🥇|6s|2014| 

### Environment: [Colab](https://colab.research.google.com/drive/1EFNDgYf4vFnK83LSGLC4FPvGPOgAH3v1)

#### without CPU
|Model| Option | Total Execution Time | Total Tokenization Time (first → last lines) |  iterations/sec |
|:---:|:--------:|:----:|:-------:|:-------:|
|AttaCut-C|--batch-size=1    --num-cores=1|1m5s🥇|59s |234| 
||--batch-size=100 --num-cores=1|1m27s|1m24s|165| 
||--batch-size=200 --num-cores=1|1m37|1m34s|147| 
|AttaCut-SC|--batch-size=1    --num-cores=1|1m14s| 36s | 381| 
||--batch-size=100 --num-cores=1|1m13s🥇|36s|378| 
||--batch-size=200 --num-cores=1|1m17s|40s|346| 

#### with GPU
|Model| Option | Total Execution Time  | Total Tokenization Time (first → last lines) |  iterations/sec |
|:---:|:--------:|:----:|:-------:|:-------:|
|AttaCut-C|--gpu --batch-size=1    --num-cores=1|31s|19s |714| 
||--gpu --batch-size=100 --num-cores=1|12s🥇|6s|2083| 
||--gpu --batch-size=200 --num-cores=1|13s|7s|1855| 
|AttaCut-SC|--gpu --batch-size=1    --num-cores=1|1m2s| 23s | 598| 
||--gpu --batch-size=100 --num-cores=1|43s🥇|5s|2780| 
||--gpu --batch-size=200 --num-cores=1|44s|5s|2734| 



### Environment: aws's m5.4xlarge

|Model| Option | Total Execution Time  | Total Tokenization Time (first → last lines) |  iterations/sec |
|:---:|:--------|:----:|:-------:|:-------:|
|AttaCut-C| --batch-size=1    --num-cores=1|50s|47s |292| 
||--batch-size=1 --num-cores=5|37s|35s🥇|397| 
||--batch-size=1000 --num-cores=5|1m8s|1m5s|212| 
||--batch-size=1 --num-cores=10|50s|47s|292| 
|AttaCut-SC|--batch-size=1    --num-cores=1|1m10s| 30s | 457| 
||--batch-size=100 --num-cores=1|1m13s|33s|414| 
||--batch-size=200 --num-cores=1|1m21s|40s|340| 
||--batch-size=1 --num-cores=10|1m9s|28s 🥇|484| 
||--batch-size=1 --num-cores=15|1m13s|33s|416|