@wannaphong 2020-03-13
I suggest you to add specify GPU usage.