@wannaphong 2020-03-13
@heytitle I have 2 GPUs indexed as 0,1. I need to use GPUs indexed 1.