@heytitle 2020-03-13
I usually set `export CUDA_VISIBLE_DEVICES=1`. If you want to use GPU-`0`, you set
`export CUDA_VISIBLE_DEVICES=0`. See also https://stackoverflow.com/a/39661999.